neris/requirements.txt
django
djongo
djangorestframework
pyyaml
requests
python-dotenv
